[coverage] Split profdata_merge_worker.py into separate files within a module
